tests/wpt/web-platform-tests/css/css-transitions/parsing/transition-property-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transitions: parsing transition-property with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-transitions/#propdef-transition-property">
<meta name="assert" content="transition-property supports only the grammar 'none | <single-transition-property> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("transition-property", 'one two three');
test_invalid_value("transition-property", '1, 2, 3');
test_invalid_value("transition-property", 'none, one');

test_invalid_value("transition-property", 'initial, top');
test_invalid_value("transition-property", 'top, initial');
test_invalid_value("transition-property", 'inherit, top');
test_invalid_value("transition-property", 'top, inherit');
test_invalid_value("transition-property", 'unset, top');
test_invalid_value("transition-property", 'top, unset');
test_invalid_value("transition-property", 'default, top');
test_invalid_value("transition-property", 'top, default');
test_invalid_value("transition-property", 'revert, top');
test_invalid_value("transition-property", 'top, revert');
test_invalid_value("transition-property", 'revert-layer, top');
test_invalid_value("transition-property", 'top, revert-layer');
</script>
</body>
</html>
